Rename the directory cache.repo to db.repo
