+++ src/com/futurewei/alcor/controller/db/ignite/IgniteConfiguration.java
@@ -1,3 +1,19 @@
+/*
+Copyright 2019 The Alcor Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+        you may not use this file except in compliance with the License.
+        You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+        Unless required by applicable law or agreed to in writing, software
+        distributed under the License is distributed on an "AS IS" BASIS,
+        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+        See the License for the specific language governing permissions and
+        limitations under the License.
+*/
+
 package com.futurewei.alcor.controller.db.ignite;
 
 import com.futurewei.alcor.controller.logging.Logger;
@@ -16,8 +32,8 @@ import org.springframework.util.Assert;
 import java.util.logging.Level;
 
 @Configuration
-@ComponentScan("com.futurewei.alcor.controller.cache")
-@EntityScan("com.futurewei.alcor.controller.cache")
+@ComponentScan("com.futurewei.alcor.controller.db")
+@EntityScan("com.futurewei.alcor.controller.db")
 @ConditionalOnProperty(prefix = "ignite", name = "host")
 public class IgniteConfiguration {
     private static final Logger logger = LoggerFactory.getLogger();
