+++ src/com/futurewei/alcor/controller/db/repo/ICacheRepository.java
@@ -14,7 +14,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
         limitations under the License.
 */
 
-package com.futurewei.alcor.controller.cache.repo;
+package com.futurewei.alcor.controller.db.repo;
 
 import com.futurewei.alcor.controller.exception.CacheException;
