+++ src/com/futurewei/alcor/controller/web/PortController.java
@@ -17,9 +17,9 @@ Licensed under the Apache License, Version 2.0 (the "License");
 package com.futurewei.alcor.controller.web;
 
 import com.futurewei.alcor.controller.app.onebox.*;
-import com.futurewei.alcor.controller.cache.repo.PortRedisRepository;
-import com.futurewei.alcor.controller.cache.repo.SubnetRedisRepository;
-import com.futurewei.alcor.controller.cache.repo.VpcRedisRepository;
+import com.futurewei.alcor.controller.db.repo.PortRedisRepository;
+import com.futurewei.alcor.controller.db.repo.SubnetRedisRepository;
+import com.futurewei.alcor.controller.db.repo.VpcRedisRepository;
 import com.futurewei.alcor.controller.exception.ParameterNullOrEmptyException;
 import com.futurewei.alcor.controller.exception.ParameterUnexpectedValueException;
 import com.futurewei.alcor.controller.exception.ResourceNotFoundException;
