+++ src/resources/application.properties
@@ -1,7 +1,7 @@
 #Machine configuration
 alcor.machine.config=classpath:.\\config\\machine.json
 #Connecton configuration
-spring.redis.host=192.168.131.131
+spring.redis.host=localhost
 spring.redis.port=6379
 apache.kafka.address=172.17.0.1:9092
 #Logging configuration
